requirements.txt
numpy>=1.24.0
opencv-python>=4.10.0
pytest>=7.4.0
pybind11>=2.11.0
ffmpeg-python==0.2.0
torch==1.10.0
torchvision==0.11.0
tqdm==4.62.0
